# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PedeAí API",
    "description": "API de gestão de pedidos para restaurantes.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/api/orders/{orderId}/marketplace/cancellation": {
      "post": {
        "tags": [
          "marketplace-order-controller"
        ],
        "operationId": "requestCancellation",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketplaceCancellationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OutboundActionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/marketplace-actions/{id}/retry": {
      "post": {
        "tags": [
          "marketplace-order-controller"
        ],
        "operationId": "retry_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OutboundActionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/integrations": {
      "get": {
        "tags": [
          "integration-controller"
        ],
        "operationId": "list_7",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConnectionResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "integration-controller"
        ],
        "operationId": "connect",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/integrations/{id}/simulated-orders": {
      "post": {
        "tags": [
          "integration-controller"
        ],
        "operationId": "simulate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/integrations/{id}": {
      "patch": {
        "tags": [
          "integration-controller"
        ],
        "operationId": "update_10",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/orders/{orderId}/marketplace/sync": {
      "get": {
        "tags": [
          "marketplace-order-controller"
        ],
        "operationId": "sync",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OutboundActionResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{orderId}/marketplace/cancellation-reasons": {
      "get": {
        "tags": [
          "marketplace-order-controller"
        ],
        "operationId": "cancellationReasons",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CancellationReasonResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MarketplaceCancellationRequest": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 20,
            "minLength": 0
          },
          "description": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          }
        },
        "required": [
          "code",
          "description"
        ]
      },
      "OutboundActionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "action": {
            "type": "string",
            "enum": [
              "CONFIRM",
              "START_PREPARATION",
              "READY",
              "DISPATCH",
              "REQUEST_CANCELLATION"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "DONE",
              "FAILED",
              "SKIPPED"
            ]
          },
          "attempts": {
            "type": "integer",
            "format": "int32"
          },
          "lastError": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "action",
          "attempts",
          "createdAt",
          "id",
          "lastError",
          "status"
        ]
      },
      "ConnectionRequest": {
        "type": "object",
        "properties": {
          "externalMerchantId": {
            "type": "string",
            "maxLength": 80,
            "minLength": 0
          },
          "autoConfirm": {
            "type": "boolean"
          }
        },
        "required": [
          "autoConfirm",
          "externalMerchantId"
        ]
      },
      "ConnectionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "provider": {
            "type": "string",
            "enum": [
              "PEDEAI",
              "IFOOD",
              "NINETY_NINE_FOOD"
            ]
          },
          "externalMerchantId": {
            "type": "string"
          },
          "merchantName": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "PAUSED",
              "ERROR"
            ]
          },
          "autoConfirm": {
            "type": "boolean"
          },
          "lastEventAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "lastError": {
            "type": [
              "string",
              "null"
            ]
          },
          "failedActions": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "autoConfirm",
          "externalMerchantId",
          "failedActions",
          "id",
          "lastError",
          "lastEventAt",
          "merchantName",
          "provider",
          "status"
        ]
      },
...
      "ConnectionUpdateRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "PAUSED",
              "ERROR"
            ]
          },
          "autoConfirm": {
            "type": "boolean"
          }
        },
        "required": [
          "autoConfirm",
          "status"
        ]
      },
...
      "CancellationReasonResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "description"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}